move NewAppleSignupView into binders package
